prep work and tests
