+++ introspection-engine/connectors/sql-introspection-connector/src/misc_helpers.rs
@@ -265,6 +265,9 @@ pub(crate) fn calculate_backrelation_field(
 }
 
 pub(crate) fn calculate_default(table: &Table, column: &Column, arity: &FieldArity) -> Option<DefaultValue> {
+    println!("DEFAULT VALUE{:?}", column.default);
+
+    //todo make cases with default value explicit for every datatype??
     match (&column.default, &column.tpe.family) {
         (_, _) if *arity == FieldArity::List => None,
         (Some(d), ColumnTypeFamily::Boolean) => match parse_int(d) {
@@ -279,6 +282,7 @@ pub(crate) fn calculate_default(table: &Table, column: &Column, arity: &FieldAri
         (Some(d), ColumnTypeFamily::Float) => parse_float(d).map(|x| DefaultValue::Single(ScalarValue::Float(x))),
         (Some(d), ColumnTypeFamily::String) => Some(DefaultValue::Single(ScalarValue::String(d.to_string()))),
         (Some(_), ColumnTypeFamily::DateTime) => None, //todo
+        (Some(d), ColumnTypeFamily::Enum(_)) => None,  //todo
         (None, _) if column.auto_increment => Some(DefaultValue::Expression(ValueGenerator::new_autoincrement())),
         (_, _) => None,
     }
