+++ introspection-engine/connectors/sql-introspection-connector/tests/db_specific_introspection/mysql/mod.rs
@@ -1,3 +1,4 @@
+pub mod enums_mysql;
 pub mod relations_mysql;
 pub mod relations_with_compound_fk_mysql;
 pub mod remapping_database_names_mysql;
