+++ introspection-engine/connectors/sql-introspection-connector/tests/db_specific_introspection/mysql/tables_mysql.rs
@@ -143,28 +143,32 @@ async fn introspecting_a_table_with_required_and_optional_columns_must_work(api:
     custom_assert(&result, dm);
 }
 
-//#[test_each_connector(tags("mysql"))]
-//#[ignore]
-//fn introspecting_a_table_with_datetime_default_values_should_work(api: &TestApi) {
-//    let barrel = api.barrel();
-//    let _setup_schema = barrel.execute_with_schema(, api.db_name(|migration| ){
-//        migration.create_table("User", |t| {
-//            t.add_column("id", types::primary());
-//            t.add_column("name", types::text());
-//            t.inject_custom("`joined` date DEFAULT CURRENT_DATE")
-//        });
-//    }).await;
+// #[test_each_connector(tags("mysql"))]
+// async fn introspecting_a_table_with_datetime_default_values_should_work(api: &TestApi) {
+//     let barrel = api.barrel();
+//     let _setup_schema = barrel
+//         .execute_with_schema(
+//             |migration| {
+//                 migration.create_table("User", |t| {
+//                     t.add_column("id", types::primary());
+//                     t.add_column("name", types::text());
+//                     t.inject_custom("`joined` date DEFAULT CURRENT_DATE")
+//                 });
+//             },
+//             api.db_name(),
+//         )
+//         .await;
 //
-//    let dm = r#"
-//            model User {
-//                id      Int @id
-//                joined DateTime? @default(now())
-//                name String
-//            }
-//        "#;
-//    let result = dbg!(api.introspect().await);
-//    custom_assert(&result, dm);
-//}
+//     let dm = r#"
+//             model User {
+//                 id      Int @id
+//                 joined DateTime? @default(now())
+//                 name String
+//             }
+//         "#;
+//     let result = dbg!(api.introspect().await);
+//     custom_assert(&result, dm);
+// }
 
 #[test_each_connector(tags("mysql"))]
 async fn introspecting_a_table_with_default_values_should_work(api: &TestApi) {
@@ -255,90 +259,6 @@ async fn introspecting_a_table_with_a_multi_column_non_unique_index_should_work(
         "#;
     let result = dbg!(api.introspect().await);
     custom_assert(&result, dm);
-}
-
-#[test_each_connector(tags("mysql"))]
-async fn introspecting_a_table_enums_should_work(api: &TestApi) {
-    api.barrel()
-        .execute(|migration| {
-            migration.create_table("Book", |t| {
-                t.add_column("id", types::primary());
-                t.inject_custom("color  ENUM ( 'black', 'white') Not Null");
-                t.inject_custom("color2  ENUM ( 'black2', 'white2') Not Null");
-            });
-        })
-        .await;
-
-    let dm = r#"
-        model Book {
-            color   Book_color
-            color2  Book_color2
-            id      Int     @default(autoincrement()) @id
-        }
-
-        enum Book_color{
-            black
-            white
-        }
-
-        enum Book_color2{
-            black2
-            white2
-        }
-    "#;
-
-    let result = dbg!(api.introspect().await);
-    let result1 = dbg!(api.introspect().await);
-    let result2 = dbg!(api.introspect().await);
-    let result3 = dbg!(api.introspect().await);
-    let result4 = dbg!(api.introspect().await);
-    custom_assert(&result, dm);
-    custom_assert(&result1, dm);
-    custom_assert(&result2, dm);
-    custom_assert(&result3, dm);
-    custom_assert(&result4, dm);
-}
-
-#[test_each_connector(tags("mysql"))]
-async fn introspecting_a_table_enums_should_return_alphabetically_even_when_in_different_order(api: &TestApi) {
-    api.barrel()
-        .execute(|migration| {
-            migration.create_table("Book", |t| {
-                t.add_column("id", types::primary());
-                t.inject_custom("color2  ENUM ( 'black2', 'white2') Not Null");
-                t.inject_custom("color  ENUM ( 'black', 'white') Not Null");
-            });
-        })
-        .await;
-
-    let dm = r#"
-        model Book {
-            color   Book_color
-            color2  Book_color2
-            id      Int     @default(autoincrement()) @id
-        }
-
-        enum Book_color{
-            black
-            white
-        }
-
-        enum Book_color2{
-            black2
-            white2
-        }
-    "#;
-
-    let result = dbg!(api.introspect().await);
-    let result1 = dbg!(api.introspect().await);
-    let result2 = dbg!(api.introspect().await);
-    let result3 = dbg!(api.introspect().await);
-    let result4 = dbg!(api.introspect().await);
-    custom_assert(&result, dm);
-    custom_assert(&result1, dm);
-    custom_assert(&result2, dm);
-    custom_assert(&result3, dm);
-    custom_assert(&result4, dm);
 }
 
 #[test_each_connector(tags("mysql"))]
